tts-service/requirements.txt
flask
flask-cors
python-dotenv
boto3
python-dotenv
gtts
google-cloud-texttospeech
ffmpeg-python
